merge: Add pdonlon's code to Android, create test suite
This patch adds the Present, Enemy, and Sack classes from pdonlon's last
commit into the Android code of the application.

A JUnit test suite has also been created to test Sack and Enemy, but
still needs to be filled out. An example for testAdd() for Present is
included.

A bunch of TODOs have also been added in the merged code.

Signed-off-by: Ricky Ayoub <vgmoose@gmail.com>